// generated/src/app/api-models/ar-cash-receipt.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ArCashReceipt {
  seqCashReceipt: number;
  customerType: string;
  customerId: string;
  transDate: Date;
  transNo: string;
  transReceiptDate: Date;
  transactionType: string;
  companyCode: string;
  glRefCode: string;
  transAmt: number;
  subscriberId: string;
  description: string;
  seqCashBatchId: number;
  statementStatus: string;
  invoiceNo: number;
  postDate: Date;
  seqGpbilId: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}